코드 로직 거의 완료... 중간 저장

+++ Engine/Resource/Model3D/FBXLoader.h
@@ -50,7 +50,7 @@ namespace ehw
 		std::vector<float3>						vecBinormal{};
 		std::vector<float3>						vecTangent{};
 
-		std::vector<float4>						vecBlendIndex{};
+		std::vector<uint4>						vecBlendIndex{};
 		std::vector<float4>						vecBlendWeight{};
 
 		std::vector<std::vector<UINT>>			vecIndexBuffers{};
